Reduce orphans instances
